layout/base/tests/chrome/dialog_with_positioning_window.xhtml

<?xml version="1.0" encoding="UTF-8" ?>
<?xml-stylesheet href="chrome://global/skin/global.css" type="text/css"?>
<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="setTimeout(runTest, 0)">
  <vbox>
    <label value="powered by example.com" style="padding: 16px;"/>
  </vbox>
  <hbox id="t" style="display: block; position: fixed; right: 16px; bottom: 16px;">
    <button label="OK"/>
  </hbox>
<script><![CDATA[
var SimpleTest = window.arguments[0].SimpleTest;
var SpecialPowers = window.arguments[0].SpecialPowers;
var is = window.arguments[0].is;
var ok = window.arguments[0].ok;

// We run this off a setTimeout from onload, because the XUL window
// only does its intrinsic-height layout after the load event has
// finished
function runTest() {
  var t = document.getElementById("t");
  var tBottom = t.getBoundingClientRect().bottom;
  is(tBottom, document.documentElement.getBoundingClientRect().bottom - 16,
     "check fixed-pos element t bottom positioned correctly");
  ok(tBottom < 200, "fixed-pos element t bottom must be sane, less than 200 (got " + tBottom + ")");
  window.close();
  SimpleTest.finish();
}
]]></script>
</window>
